% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runcomments.r
\name{getSelectionText}
\alias{getSelectionText}
\title{Return text from current source editor selection}
\usage{
getSelectionText()
}
\value{
a character vector. One element per selected line
}
\description{
Return text from current source editor selection
}
