% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noComment.r
\name{noComment}
\alias{noComment}
\title{Run selected, commented lines from RStudio source editor in the console}
\usage{
noComment(pattern, ...)
}
\arguments{
\item{pattern}{character string containing a regular expression}

\item{...}{other arguments to rstudioapi::sendToConsole}
}
\value{
nothing
}
\description{
Run selected, commented lines from RStudio source editor in the console
}
