% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noComment.r
\name{removeCommentCharacters}
\alias{removeCommentCharacters}
\title{Use regex pattern to remove comment characters from text}
\usage{
removeCommentCharacters(text, pattern = NULL)
}
\arguments{
\item{text}{character vector of text from current source editor selection}

\item{pattern}{character string containing a regular expression}
}
\value{
character vector with comment characters removed
}
\description{
Use regex pattern to remove comment characters from text
}
