% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runcomments.r
\name{runCommentedSource}
\alias{runCommentedSource}
\title{Run selected, commented lines from RStudio source editor in the console}
\usage{
runCommentedSource(pattern, ...)
}
\arguments{
\item{pattern}{character string containing a regular expression}

\item{...}{other arguments to rstudioapi::sendToConsole}
}
\value{

}
\description{
Run selected, commented lines from RStudio source editor in the console
}
